Replacing tablegen file
The tablegen file that comes with LLVM doesn't quite do it for me.
Things like the spell-checker are a little overly gumptious, checking
the identifier names, which isn't something I want. I still want the
checker operating on strings and comments.

I also wanted to see the colouring behaving more like the C++ syntax
highlighting.

Lastly, mine does block folding so that block definitions, bloc class
descriptions, and code blocks are automatically foldable.

I'm not sure that I'll be keeping the colors exactly as they are, but
I'm happier with how it is now than how it was.

Still todo:
 - Hex/Binary colouring isn't working
 - No block comments