Update rofi config
Rofi 1.7 stopped supporting the Xresource style rofi configuration file.
I haven't touched this config in quite some time, so I decided to spruce
it up a bit.

I've also been using a mac more recently, so the mod+d was getting
burdensome. I've switched it to match the mac, so it's not mod+space.