Add blocks for iface and memory
Adding implementations for network interface and memory usage blocks.
The networking interface pulls out whether the interface is enabled, and
if it is, shows the IP address. If it is disabled, it will say that it's
down.

The memory block calls free and extracts out the total and used memory
and presents that as a fraction in terms of gigabytes.