Add mega lua bits to nvim
So this makes some pretty big changes to how my nvim config works.

As a test, I've pulled the keybindings into lua to make sure that I had
everything set up right. That worked nicely and actually looks pretty
good.

The second bit updates the status line to include the filemode and info
about the number of changes to a file and the branch that the git repo
is currently on, if there is a repo. The repo head info is pulled from a
call to vim-fugitive, and the number of additions, changed lines, and
removed lines comes from a buffer-variable 'b:gitstats' which is a list
of three numbers containing the number of lines added, lines changed,
and lines removed. vim-signify computes that information readily, but
doesn't expose it, so adding
'let b:gitstats = [added, modified, deleted]` at the bottom of
`sy#sign#process_diff` will expose those stats to the status bar.