Add more LSP!
This patch adds more LSP. On macOS, we have sourcekit-lsp, which will
farm out to clangd, but clangd doesn't seem to be directly accessible.
If we have `clangd` elsewhere, we should use that directly for c++
files.

Note that `def` files are a bit different. They tend to contain a bunch
of x-macros and aren't really a complete file by themselves so we don't
really want LSP trying to figure them out.