Little cleanups
Removing a bunch of unnecessary packages from the list. With all the LSP
stuff going in, they aren't necessary anymore. Deoplete works really
well on single-file programs, but seems to tip over with things at the
scale of LLVM/Swift. When it did provide results, there were far too
many to be useful and were pretty rarely what I was looking for.

I'm hoping to figure out how to get the LSP stuff to do decent
completion, but for the time being, definition lookup will be the
easiest way to go.

Also cleaning up little things here and there.