Remove ARCHFLAGS environment variable
I pulled this in from some time in university when I was working with
the build system for some library that needed it set. It worked then
because all of my machines were Intel. This has a really bad interaction
with pip/setuptools in python when setting up libraries on aarch64
machines. On the Mac, you end up with a thin x86_64 library, which is
definitely not the right thing. With it gone, we end up with the fat
binary with both the x86_64 and arm64 slice.