Remove LSP auto formatting
It was an interesting idea to have LSP auto format documents, but ends
up breaking things more frequently than I really wanted. The issue with
working with existing codebases is that they aren't perfectly formatted
from the get-go, so this would reformat the entire file, which lead to
massive unrelated patches.