Move to makefile installation
compnerd suggested moving to a makefile for handling dotfile
installation. I have to agree, it's nice, so we're moving to that
instead of a custom script.

With this, it's rather easy to set the set of components to install for
each platform, or to add configurations for a single tool/component.
Right now we have targets for these tools:
 - git
 - i3
 - lldb
 - nvim
 - rofi
 - tmux
 - zsh

I don't have automated installation for the terminal emulator profiles.