[git] Enable ssh-key commit signing
Git 2.34 added ssh-based commit signing, which is extra dope.
Setting it up now. There is one, per-machine, change that needs to be
made. The config file needs to include the signkey, which is obtainable
by running `cat` on the `id_<algo>.pub` file and pasting that into the
config file.

Secondly, the `authorized_signatures` list must be in the format:
`<email>[,<email>...] <key type> <public key>`

Note that the format accepts wildcards for the email portion, so you can
use `*` to accept any email address. The key type is the same as
whatever ssh is saying, so it can be `ssh-ed25519`, or `ssh-rsa`, or
whatever is in the public key file.

You may need to add the ssh key to the ssh-agent for the local checking
to work right.

Something like the following should do the trick:
```
Host *
  AddKeysToAgent yet
  UseKeychain yes
  IdentityFile ~/.ssh/id_ed25519
```