[Git]: Add activate/deactivate git aliases
There are times when I want to ensure that I'm not pushing to a specific
remote, but also want to be able to enable pushing to that remote again.

The 'activate' alias sets the PUSH URL for the
specified to remote to the PULL URL for that remote.

The 'deactivate' alias replaces the PUSH URL for the specified remote to
'INVALID', which will ensure that pushes will fail.