vim: vim-specific configs
Adding a vim-specific configuration so that the nvim configuration can
migrate to lua, while still generally giving me the behavior I'm used
to.

This is a super trimmed down configuration. There aren't any plugins
because I need to move off of vundle and don't think I need the plugins
for limited use.