[nvim] wildmenu/wildmode config
Fixing the wildmenu configuration. Was defaulting to the first match,
which in some cases was long and unpleasant, resulting in lots of
backspacing and retyping. 'list:longest' lists all of the options, and
only completes up to the first character that is not consistent instead
of just choosing the first.