[nvim] clean up line numbers
Switching to have both rnu and number set instead of flipping between
the two depending on whether working in normal mode vs edit mode. It's a
fragment from when I was working with a vim that didn't support having
both on at the same time, but is not necessary in nvim, so removing it
and the autocommands associated with it.