[git] Trimming configuration
Removing the private bits of the git configuration. All of the key
signature stuff is causing trouble since it's specific to each machine,
so updating my dotfiles and rebasing has been a bit of a headache.
Pulling that out into a separate file that isn't tracked but should sit
next to the checkout.