[nvim] Swift SourceKit-lsp reload lsp on change
nvim 0.9.0 seems to have the capacity for handling dynamic registration
for 'didChangeWatchedFiles', so when `compile_commands.json` updates, we
can reload it without having to close and re-open the editor. I'm only
enabling this for sourcekit-lsp since that seems to rely more heavily on
it than the other servers and this seems like a pretty sketchy fix.