fzf: Update default command
The default fzf command would often ignore files and directories under a
directory with a capital first letter. This is bad when working on
projects like CMake, which have `Modules/`, `Help/`, `Source/` and
`Tests/`. This update uses ripgrep to drive the search if it's
installed. Otherwise we don't override the command.